Move to other file, add comments and a warning for unknown params
